Replace puppeteer implementation with playwright. Move files over. Adjust Dockerfile.
