+++ lib/chromiumBrowser.js
@@ -1,68 +1,53 @@
-const proxyChain = require('proxy-chain')
-const puppeteer = require('puppeteer-core')
-const { resolveTimeoutWithNeedIncrease } = require('./timeoutHelper')
+const proxyChain = require('proxy-chain');
+const { chromium } = require('playwright');
+const logger = require('./logger');
+
+const BASE_ARGS = [
+  '--no-sandbox',
+  '--disable-dev-shm-usage',
+  '--disable-blink-features=AutomationControlled',
+  '--disable-features=IsolateOrigins,site-per-process',
+  '--disable-font-subpixel-positioning',
+];
 
 class ChromiumBrowser {
-  args = [
-    '--no-sandbox',
-    '--disable-setuid-sandbox',
-    '--disable-web-security',
-    '--disable-features=IsolateOrigins,site-per-process',
-    '--disable-features=TranslateUI',
-    '--disable-extensions',
-    '--disable-background-timer-throttling',
-    '--disable-renderer-backgrounding',
-    '--disable-dev-shm-usage',
-    '--disable-client-side-phishing-detection',
-    '--ignore-certificate-errors',
-    '--js-flags=--max-old-space-size=2048',
-    '--autoplay-policy=user-gesture-required',
-    '--disable-font-subpixel-positioning'
-  ];
-
-  browser = null
-  debug = false
-  local = false
-  anonymizedProxy = null
-
-  localExecutivePath = '/usr/bin/chromium-browser';
-
   constructor(debug = false, local = false) {
     this.debug = debug;
     this.local = local;
+    this.browser = null;
+    this.anonymizedProxy = null;
+    this.staticArgs = BASE_ARGS;
+    this.localExecutivePath = '/usr/bin/chromium-browser';
   }
 
-  /**
-   * Get browser instance.
-   * @return {Promise<Browser>}
-   */
-  async getBrowser (proxy, options = {}) {
-    const launchArgs = [...this.args];
-    const needIncrease = options?.needIncrease;
+  async getBrowser(proxy) {
+    const launchArgs = [...this.staticArgs];
 
-    const protocolTimeout = resolveTimeoutWithNeedIncrease(needIncrease, 180000);
+    const launchOptions = {
+      args: launchArgs,
+      headless: true,
+      chromiumSandbox: this.local,
+      timeout: 120000,
+    };
 
     if (proxy) {
       this.anonymizedProxy = await proxyChain.anonymizeProxy(proxy);
       launchArgs.push(`--proxy-server=${this.anonymizedProxy}`);
     }
 
-    try {
-      return puppeteer.launch({
-        args: launchArgs,
-        protocolTimeout,
-        defaultViewport: { width: 800, height: 600},
-        executablePath: this.localExecutivePath,
-        headless: 'new',
-        dumpio: false,
-        ignoreHTTPSErrors: true,
-      })
-    } catch (e) {
-      console.log('Error launching browser:', e);
+    let executablePath = this.localExecutivePath;
+
+    if (executablePath) {
+      launchOptions.executablePath = executablePath;
     }
+
+    if (this.debug) logger.debug('Launching Chromium with options', { launchOptions });
+
+    this.browser = await chromium.launch(launchOptions);
+    return this.browser;
   }
 
-  async closeProxy () {
+  async closeProxy() {
     if (this.anonymizedProxy) {
       await proxyChain.closeAnonymizedProxy(this.anonymizedProxy, true);
       this.anonymizedProxy = null;
