+++ lib/func.js
@@ -6,7 +6,109 @@ const url = require('url')
 const crypto = require('crypto')
 const sharp = require('sharp')
 const logger = require('./logger')
-const { resolveTimeoutWithNeedIncrease } = require('./timeoutHelper')
+
+const FONT_HOST_DENYLIST = new Set(['fonts.gstatic.com'])
+const FONT_FILE_PATTERN = /\.(woff2?|ttf|otf)(?:[?#].*)?$/i
+const PANTHEON_HOST_PATTERN = /(^|\.)pantheonsite\.io$/i
+
+const parseHostname = (input) => {
+  if (!input || typeof input !== 'string') {
+    return ''
+  }
+  try {
+    return new URL(input).hostname.toLowerCase()
+  } catch (_) {
+    return ''
+  }
+}
+
+const isFontLikeRequest = (requestUrl, resourceType = '') => {
+  if (!requestUrl) {
+    return resourceType === 'font'
+  }
+  const hostname = parseHostname(requestUrl)
+  if (hostname && FONT_HOST_DENYLIST.has(hostname)) {
+    return true
+  }
+
+  if (resourceType === 'font') {
+    return true
+  }
+
+  try {
+    const pathname = new URL(requestUrl).pathname || ''
+    return FONT_FILE_PATTERN.test(pathname)
+  } catch (_) {
+    return FONT_FILE_PATTERN.test(requestUrl)
+  }
+}
+
+const shouldEnableDeterrenceBypass = (jobUrl) => {
+  if (!jobUrl) {
+    return false
+  }
+  const hostname = parseHostname(jobUrl)
+  return Boolean(hostname && PANTHEON_HOST_PATTERN.test(hostname))
+}
+
+const shouldApplyDeterrenceHeader = (deterrenceState, requestUrl, resourceType = '') => {
+  if (!deterrenceState?.enabled || !requestUrl) {
+    return false
+  }
+
+  if (isFontLikeRequest(requestUrl, resourceType)) {
+    return false
+  }
+
+  const hostname = parseHostname(requestUrl)
+  if (!hostname) {
+    return false
+  }
+
+  return PANTHEON_HOST_PATTERN.test(hostname)
+}
+
+const describeError = (err) => {
+  if (!err) {
+    return 'unknown error'
+  }
+  if (Object.hasOwn(err, 'message') && err.message) {
+    return err.message
+  }
+  return String(err)
+}
+
+const isTargetClosureError = (err) => {
+  if (!err) {
+    return false
+  }
+  const message = describeError(err)
+  return /Target closed|Session closed|Page closed|Browser has been closed|Protocol error/i.test(message)
+}
+
+const ensurePageOpen = (page, label = 'operation') => {
+  if (!page || (typeof page.isClosed === 'function' && page.isClosed())) {
+    throw new Error(`Target closed before ${label}`)
+  }
+}
+
+const pickUserAgentFromHeaders = (job) => {
+  if (!checkArgs(job, 'headers', true)) {
+    return ''
+  }
+
+  const userAgentHeader = job.args.headers.filter(item => {
+    return (Object.hasOwn(item, 'header') && item.header && item.header.toLowerCase() === 'user-agent')
+  })
+
+  if (userAgentHeader && userAgentHeader.length) {
+    if (Object.hasOwn(userAgentHeader[0], 'value') && userAgentHeader[0].value.length) {
+      return userAgentHeader[0].value
+    }
+  }
+
+  return ''
+}
 
 const checkArgs = (obj, field, checkLength = false) => {
   let result = (Object.hasOwn(obj, 'args') && obj.args && Object.hasOwn(obj.args, field))
@@ -18,21 +120,92 @@ const checkArgs = (obj, field, checkLength = false) => {
 }
 
 const updatePageViewport = async (page, job, maxPageHeight = null) => {
-  let scrollHeight = await page.evaluate(`(async () => {
-        return document.documentElement.scrollHeight;
-    })()`)
+  ensurePageOpen(page, 'updatePageViewport evaluate')
+
+  let scrollHeight
+  try {
+    scrollHeight = await page.evaluate(() => document.documentElement.scrollHeight)
+  } catch (err) {
+    if (isTargetClosureError(err)) {
+      throw new Error(`Target closed during updatePageViewport evaluate: ${describeError(err)}`)
+    }
+    throw err
+  }
 
   if (maxPageHeight && scrollHeight > maxPageHeight) {
     scrollHeight = maxPageHeight
   }
 
-  await page.setViewport({ width: Number.parseInt(job.breakpoint), height: Number.parseInt(scrollHeight) })
-  await new Promise(resolve => setTimeout(resolve, 1000));
+  ensurePageOpen(page, 'updatePageViewport resize')
+  try {
+    await page.setViewportSize({
+      width: Number.parseInt(job.breakpoint, 10),
+      height: Number.parseInt(scrollHeight, 10)
+    })
+    // Allow Chromium to finish reflowing before we continue manipulating the page
+    await page.waitForTimeout(1000)
+  } catch (err) {
+    if (isTargetClosureError(err)) {
+      throw new Error(`Target closed during updatePageViewport resize: ${describeError(err)}`)
+    }
+    throw err
+  }
+
   return scrollHeight
 }
 
 const random = (low, high) => {
   return crypto.randomInt(low, high + 1);
+}
+
+const handleCloudflareChallenge = async (page, { maxAttempts = 3, frameWaitMs = 5000, retryDelayMs = 2000 } = {}) => {
+  if (!page) {
+    return false
+  }
+
+  const challengePresent = async () => {
+    return await page.evaluate(() => {
+      const iframe = document.querySelector('iframe[src*="challenges.cloudflare.com"]')
+      const bodyText = document.body?.innerText || ''
+      const marker = bodyText.includes('Please unblock challenges.cloudflare.com')
+      const turnstile = document.querySelector('[data-cf-challenge], .cf-turnstile, #challenge-stage')
+      return Boolean(iframe || marker || turnstile)
+    })
+  }
+
+  for (let attempt = 0; attempt < maxAttempts; attempt++) {
+    const detected = await challengePresent().catch(() => false)
+    if (!detected) {
+      return false
+    }
+
+    const frameHandle = await page.$('iframe[src*="challenges.cloudflare.com"]')
+    if (frameHandle) {
+      const frame = await frameHandle.contentFrame()
+      if (frame) {
+        const button = await frame.waitForSelector('input[type="button"], button, #cf-stage input[type="submit"]', { timeout: frameWaitMs }).catch(() => null)
+        if (button) {
+          await button.click().catch(() => {})
+          await frame.waitForTimeout(1500)
+        }
+      }
+    }
+
+    await page.waitForTimeout(retryDelayMs)
+
+    if (!(await challengePresent().catch(() => false))) {
+      return true
+    }
+
+    await page.reload({ waitUntil: 'domcontentloaded', timeout: frameWaitMs * 3 }).catch(() => {})
+    await page.waitForTimeout(retryDelayMs)
+
+    if (!(await challengePresent().catch(() => false))) {
+      return true
+    }
+  }
+
+  return false
 }
 
 const awaitResponse = async (page, timeout = 60000) => {
@@ -59,6 +232,41 @@ const awaitResponse = async (page, timeout = 60000) => {
   ])
 }
 
+const USER_AGENT = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/123.0.0.0 Safari/537.36'
+
+const buildHeaderState = (job) => {
+  let userAgentString = pickUserAgentFromHeaders(job)
+
+  if (!userAgentString.length) {
+    userAgentString = USER_AGENT
+  }
+
+  const headers = {}
+
+  if (job?.args?.headers) {
+    job.args.headers.forEach(element => {
+      if (element.header && element.header.trim().length) {
+        headers[element.header] = element.value
+      }
+    })
+  }
+
+  const deterrenceEnabled = shouldEnableDeterrenceBypass(job?.url)
+  if (deterrenceEnabled) {
+    logger.debug('Deterrence-Bypass enabled for pantheonsite.io requests')
+  }
+
+  return {
+    userAgentString,
+    headers,
+    deterrenceBypass: {
+      enabled: deterrenceEnabled,
+      headerName: 'Deterrence-Bypass',
+      headerValue: '1'
+    }
+  }
+}
+
 module.exports = {
   checkArgs: (obj, field, checkLength = false) => {
     return checkArgs(obj, field, checkLength)
@@ -73,20 +281,19 @@ module.exports = {
     let totalHeight = 0;
 
     do {
-      // Need to wait if page is reloaded (see: https://github.com/ygerasimov/diffy-pm/issues/122)
-      await page.waitForSelector('body')
+      await page.waitForSelector('body', { state: 'attached' });
       scrollHeight = await page.evaluate('document.body.scrollHeight');
 
-      await page.waitForSelector('body')
+      await page.waitForSelector('body', { state: 'attached' });
       await page.evaluate('window.scrollBy(0, 100)');
       totalHeight += 100;
 
-      await new Promise(resolve => setTimeout(resolve, 100));
+      await page.waitForTimeout(100);
     } while (totalHeight < scrollHeight)
 
     try {
-      await page.evaluate('window.scrollTo(0, 0)')
-      await new Promise(resolve => setTimeout(resolve, 500));
+      await page.evaluate('window.scrollTo(0, 0)');
+      await page.waitForTimeout(500);
     } catch (e) {}
 
     return Promise.resolve();
@@ -129,10 +336,10 @@ module.exports = {
     try {
       await page.evaluate(job.args.js_code)
     } catch (e) {
-      logger.warn('Failed to evaluate page', { error: e, js_code: job.args.js_code })
+      logger.debug('Failed to evaluate page', { error: e, js_code: job.args.js_code })
     }
 
-    await new Promise(resolve => setTimeout(resolve, 2000));
+    await page.waitForTimeout(2000)
     return page
   },
 
@@ -142,13 +349,13 @@ module.exports = {
     }
 
     try {
-      await page.addStyleTag({ content: job.args.css_code })
+      await page.addStyleTag({ content: job.args.css_code });
     } catch (e) {
-      logger.error('Failed to add style tag', { error: e, css_code: job.args.css_code })
+      logger.error('Failed to add style tag', { error: e, css_code: job.args.css_code });
     }
 
-    await new Promise(resolve => setTimeout(resolve, 2000));
-    return page
+    await page.waitForTimeout(2000)
+    return page;
   },
 
   addFixtures: async (page, job) => {
@@ -158,46 +365,66 @@ module.exports = {
 
     await page.evaluate((_fixtures) => {
 
+      const clampPositive = (value, fallback) => {
+        const number = Number.parseFloat(value)
+        if (Number.isFinite(number) && number > 0) {
+          return Math.round(number)
+        }
+        return fallback
+      }
+
+      const buildInlinePlaceholder = (width, height, label = 'Diffy fixture') => {
+        const safeWidth = clampPositive(width, 300)
+        const safeHeight = clampPositive(height, 200)
+        const text = `${safeWidth}x${safeHeight}`
+        const svg = `<?xml version="1.0" encoding="UTF-8"?>\n` +
+          `<svg xmlns="http://www.w3.org/2000/svg" width="${safeWidth}" height="${safeHeight}" viewBox="0 0 ${safeWidth} ${safeHeight}">` +
+          `<defs><style>@font-face{font-family:'Inter';src:local('Arial')}</style></defs>` +
+          `<rect width="100%" height="100%" fill="#d8d8d8"/>` +
+          `<line x1="0" y1="0" x2="${safeWidth}" y2="${safeHeight}" stroke="#b0b0b0" stroke-width="2"/>` +
+          `<line x1="${safeWidth}" y1="0" x2="0" y2="${safeHeight}" stroke="#b0b0b0" stroke-width="2"/>` +
+          `<text x="50%" y="50%" dominant-baseline="middle" text-anchor="middle" font-family="'Inter',sans-serif" font-size="${Math.max(12, Math.round(Math.min(safeWidth, safeHeight) / 8))}" fill="#6b6b6b">${text}</text>` +
+          `<title>${label}</title>` +
+          `</svg>`
+
+        return `data:image/svg+xml;charset=utf-8,${encodeURIComponent(svg)}`
+      }
+
       function diffyImageFixture (el, selector) {
-        return new Promise((resolve, reject) => {
+        return new Promise((resolve) => {
           try {
-            const w = el.width || null
-            const h = el.height || null
+            const w = clampPositive(el.width || el.naturalWidth, 300)
+            const h = clampPositive(el.height || el.naturalHeight, 200)
             const src = el.src || null
 
             if (src && w && h) {
-              el.addEventListener('load', () => {
-                resolve();
-              });
-              el.addEventListener('error', (e) => {
-                // console.error('Failed to diffy image fixture', e) // TODO: Prettify error dump
-                reject(e);
-              });
+              const placeholderSrc = buildInlinePlaceholder(w, h, 'Diffy image fixture')
 
-              // @TODO add timeout in case image is not loaded
-
-              /**
-               * @TODO check if we want to depend on picsum.photos service
-               * idea: copy images for all resolutions to s3 and expose via cloudfront (fast and stable)
-               */
-
-              el.src = `https://picsum.photos/id/0/${w}/${h}`
+              try {
+                el.src = placeholderSrc
+              } catch (_) {}
 
               if (el.hasAttribute('data-src')) {
-                el.setAttribute('data-src', el.src)
+                el.setAttribute('data-src', placeholderSrc)
               }
 
               if (el.hasAttribute('srcset')) {
-                el.setAttribute('srcset', el.src + ' 1x')
+                el.setAttribute('srcset', `${placeholderSrc} 1x`)
               }
-            } else {
-              // console.error('Can\'t add diffy image fixture', selector, src, h, w) // TODO: Prettify error dump
+
+              try {
+                el.setAttribute('data-diffy-fixture', 'image')
+                el.setAttribute('data-diffy-fixture-size', `${w}x${h}`)
+              } catch (_) {}
+
+              if (typeof el.decode === 'function') {
+                el.decode().catch(() => {})
+              }
               return resolve()
             }
-          } catch (e) {
-            // console.error('Failed to diffy image fixture', e) // TODO: Prettify error dump
-            return resolve()
-          }
+          } catch (e) {}
+
+          return resolve()
         })
       }
 
@@ -205,7 +432,7 @@ module.exports = {
         return new Promise((resolve) => {
           try {
             const elStyle = el.currentStyle || window.getComputedStyle(el, false);
-            const backgroundImage = elStyle.backgroundImage.slice(4, -1).replace(/"/g, '');
+            const backgroundImage = elStyle.backgroundImage.slice(4, -1).replace(/\"/g, '');
 
             if (!backgroundImage) {
               // No background image
@@ -215,11 +442,16 @@ module.exports = {
             getImageInfo(backgroundImage)
               .then((imageInfo) => {
                 if (imageInfo.width && imageInfo.height) {
-                    const newBackgroundImageSrc = `https://picsum.photos/id/0/${Math.round(imageInfo.width)}/${Math.round(imageInfo.height)}`;
+                    const width = clampPositive(Math.round(imageInfo.width), 300);
+                    const height = clampPositive(Math.round(imageInfo.height), 200);
+                    const newBackgroundImageSrc = buildInlinePlaceholder(width, height, 'Diffy background fixture');
                     const newBackgroundImage = new Image();
                     newBackgroundImage.addEventListener('load', () => {
                         el.style.backgroundImage = 'url(' + newBackgroundImageSrc + ')';
-
+                        try {
+                          el.setAttribute('data-diffy-fixture', 'background-image');
+                          el.setAttribute('data-diffy-fixture-size', `${width}x${height}`);
+                        } catch (_) {}
                         resolve();
                     });
                     newBackgroundImage.addEventListener('error', () => {
@@ -316,24 +548,31 @@ module.exports = {
 
     return page.evaluate((_elements) => {
       const isVisible = (element) => {
-        const style = window.getComputedStyle(element);
+        if (!element) {
+          return false
+        }
+        const style = window.getComputedStyle(element)
         return (
-            style.display !== 'none' &&
-            style.visibility !== 'hidden' &&
-            style.opacity !== '0' &&
-            element.offsetWidth > 0 &&
-            element.offsetHeight > 0
-        );
-      };
+          style.display !== 'none' &&
+          style.visibility !== 'hidden' &&
+          style.opacity !== '0' &&
+          element.offsetWidth > 0 &&
+          element.offsetHeight > 0
+        )
+      }
 
-      const vrtPaintOver = (element) => {
+      const maskElement = (element) => {
         if (!isVisible(element)) {
-          return;
+          return
         }
 
-        const rect = element.getBoundingClientRect();
+        const rect = element.getBoundingClientRect()
+        if (!rect || rect.width <= 0 || rect.height <= 0) {
+          return
+        }
 
-        const overlay = document.createElement('div');
+        const overlay = document.createElement('div')
+        overlay.dataset.diffyMaskOverlay = 'true'
         Object.assign(overlay.style, {
           display: 'block',
           position: 'absolute',
@@ -341,25 +580,31 @@ module.exports = {
           left: `${rect.left + window.scrollX}px`,
           width: `${rect.width}px`,
           height: `${rect.height}px`,
-          backgroundColor: 'green',
+          backgroundColor: '#00aa00',
           zIndex: '9999',
           pointerEvents: 'none'
-        });
+        })
 
-        document.body.appendChild(overlay);
-      };
+        document.body.appendChild(overlay)
+      }
 
       window.scrollTo(0, 0)
 
-      _elements.forEach((selector) => {
-        selector = selector.trim();
-        if (!selector.length) return;
+      const selectors = Array.isArray(_elements) ? _elements : []
+      selectors.forEach((selector) => {
+        if (typeof selector !== 'string') {
+          return
+        }
+        const trimmed = selector.trim()
+        if (!trimmed.length) {
+          return
+        }
 
-        document.querySelectorAll(selector).forEach((element) => {
-          vrtPaintOver(element);
-        });
-      });
-    }, job.args.elements);
+        document.querySelectorAll(trimmed).forEach((element) => {
+          maskElement(element)
+        })
+      })
+    }, job.args.elements)
   },
 
   updatePageViewport: async (page, job, maxPageHeight = null) => {
@@ -410,20 +655,13 @@ module.exports = {
     }
 
     let url = job.args.url
-    const needIncrease = job?.args?.need_increase;
-    const timeouts = {
-      goto: resolveTimeoutWithNeedIncrease(needIncrease, 30000),
-      awaitResponse: resolveTimeoutWithNeedIncrease(needIncrease, 60000),
-      postSubmitNavigation: resolveTimeoutWithNeedIncrease(needIncrease, 120000),
-    };
 
-    // Clear cookies for url.
-    let cookies = await page.cookies(url)
-    await page.deleteCookie(...cookies)
+    const context = page.context();
+    await context.clearCookies();
 
     logger.debug(`Navigating to ${url}`);
-    await page.goto(url, { waitUntil: ['domcontentloaded', 'networkidle2'], timeout: timeouts.goto })
-    await awaitResponse(page, timeouts.awaitResponse)
+    await page.goto(url, { waitUntil: 'networkidle', timeout: 30000 })
+    await awaitResponse(page, 60000)
 
     await updatePageViewport(page, job)
 
@@ -437,8 +675,6 @@ module.exports = {
       } catch (error) {
         logger.warn('Before login css error', { error, before_login_css: job.args.before_login_css })
       }
-
-      await new Promise(resolve => setTimeout(resolve, 2000));
     }
 
     if (job.args.usernameSelector) {
@@ -460,8 +696,7 @@ module.exports = {
       // Wait for navigation to complete when either the DOM is fully loaded
       // or the network becomes idle (excluding persistent connections like WebSockets)
       // Add a 20-second delay to allow any asynchronous operations to complete before retrieving cookies
-      await page.waitForNavigation({ waitUntil: ['domcontentloaded', 'networkidle2'], timeout: timeouts.postSubmitNavigation });
-      await new Promise(resolve => setTimeout(resolve, 20000));
+      await page.waitForLoadState('networkidle', { timeout: 120000 });
       logger.debug("Navigation after login successful.");
     } catch (error) {
       logger.error('Navigation after login failed', { error })
@@ -471,7 +706,7 @@ module.exports = {
 
     logger.debug("Authentication process completed.");
 
-    return page.cookies()
+    return context.cookies()
   },
 
   removeFile: async (filepath) => {
@@ -480,6 +715,10 @@ module.exports = {
 
   random: (min, max) => {
     return random(min, max)
+  },
+
+  handleCloudflareChallenge: async (page, options = {}) => {
+    return handleCloudflareChallenge(page, options)
   },
 
   /**
@@ -499,51 +738,30 @@ module.exports = {
   //   fs.emptyDirSync(tmp)
   // },
 
-  setHeaders: async (page, job) => {
-    let userAgentString = ''
+  setHeaders: async (page, job, preparedState = null) => {
+    let state = preparedState
 
-    let headers = {}
-    if (checkArgs(job, 'headers', true)) {
-      let userAgent = job.args.headers.filter(item => {
-        return (Object.hasOwn(item, 'header') && item.header && item.header.toLowerCase() === 'user-agent')
-      })
+    if (!state) {
+      state = buildHeaderState(job)
 
-      if (userAgent && userAgent.length) {
-        if (Object.hasOwn(userAgent[0], 'value') && userAgent[0].value.length) {
-          userAgentString = userAgent[0].value
-        }
+      if (state.userAgentString.length) {
+        logger.debug('User-Agent: "' + state.userAgentString + '"')
       }
     }
 
-    if (userAgentString.length === 0) {
-      const userAgentsPool = [
-        'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/123.0.0.0 Safari/537.36'
-      ];
-
-      userAgentString = userAgentsPool[Math.floor(Math.random() * userAgentsPool.length)];
+    if (page && state.headers && Object.keys(state.headers).length) {
+      await page.setExtraHTTPHeaders(state.headers)
     }
 
-    if (userAgentString.length) {
-      logger.debug('User-Agent: "' + userAgentString + '"');
-      await page.setUserAgent(userAgentString)
-    }
+    return state
+  },
 
-    if (job.args.headers) {
-      job.args.headers.forEach(element => {
-        if (element.header.trim().length) {
-          headers[element.header] = element.value
-        }
-      })
-    }
+  buildHeaderState: (job) => {
+    return buildHeaderState(job)
+  },
 
-    if (job.url && job.url.includes('pantheonsite.io')) {
-      headers['Deterrence-Bypass'] = '1';
-      logger.debug('Deterrence-Bypass set');
-    }
-
-    if (Object.keys(headers).length) {
-      await page.setExtraHTTPHeaders(headers)
-    }
+  shouldApplyDeterrenceHeader: (state, requestUrl, resourceType = '') => {
+    return shouldApplyDeterrenceHeader(state, requestUrl, resourceType)
   },
 
   cropElement: async (page, job) => {
@@ -599,7 +817,7 @@ module.exports = {
   },
 
   getPageMhtml: async (page) => {
-    const cdp = await page.target().createCDPSession();
+    const cdp = await page.context().newCDPSession(page);
     const { data } = await cdp.send('Page.captureSnapshot', { format: 'mhtml' });
     return data;
   },
