+++ lib/funcPerform.js
@@ -1,9 +1,32 @@
-const request = require('request') // @TODO use node http module
 const uploadS3 = require('./uploadS3.js')
 const thumbnail = require('./thumbnail.js')
 const func = require('./func.js')
 const logger = require('./logger')
-const { resolveTimeoutWithNeedIncrease } = require('./timeoutHelper')
+
+const formatLoggedError = (error) => {
+  if (!error) {
+    return { message: 'Unknown error' }
+  }
+
+  if (typeof error === 'string') {
+    return { message: error }
+  }
+
+  const details = {
+    name: error.name || 'Error',
+    message: error.message || String(error)
+  }
+
+  if (error.stack) {
+    details.stack = error.stack
+  }
+
+  if (Object.hasOwn(error, 'code')) {
+    details.code = error.code
+  }
+
+  return details
+}
 
 const sendResult = (job, jobItem, data) => {
   job.status = true
@@ -24,128 +47,6 @@ const sendError = (job, error, jobItem) => {
   return job
 }
 
-// TODO Add Url checking into the process
-const checkUrl = async (url, job) => {
-  const options = {
-    method: 'HEAD',
-    rejectUnauthorized: false,
-    requestCert: false,
-    strictSSL: false,
-    insecureHTTPParser: true,
-    timeout: 20 * 1000,
-    pool: { maxSockets: Infinity },
-    headers: { 'user-agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.11; rv:46.0) Gecko/20100101 Firefox/46.0' },
-  }
-
-  if (func.checkArgs(job, 'headers', true)) {
-    job.args.headers.forEach(item => {
-      if (item.hasOwnProperty('header') && item.header) {
-        if (item.header.toLowerCase() === 'user-agent' && item.hasOwnProperty('value') && item.value.length) {
-          options.headers['user-agent'] = item.value;
-        }
-
-        if (item.header.toLowerCase() === 'x-vercel-protection-bypass') {
-          options.headers['x-vercel-protection-bypass'] = item.value;
-        }
-      }
-    });
-  }
-
-  if (
-    job.hasOwnProperty('basicAuth') && job.basicAuth &&
-    job.basicAuth.hasOwnProperty('user') && job.basicAuth.user &&
-    job.basicAuth.hasOwnProperty('password') && job.basicAuth.password
-  ) {
-    options.auth = {
-      user: job.basicAuth.user,
-      pass: job.basicAuth.password
-    }
-  }
-
-  let lastException;
-
-  try {
-    await checkURLRequest(url, options, job)
-    return true
-  } catch (e) {
-    logger.error('Failed to checkUrl', { url, error: e });
-    lastException = e;
-  }
-
-  try {
-    options.method = 'GET'
-    await checkURLRequest(url, options, job)
-    return true
-  } catch (e) {
-    logger.error('Failed to checkUrl (GET)', { url, error: e });
-    lastException = e;
-  }
-
-  if (job.args.url) {
-    try {
-      await checkURLRequest(job.args.url, options, job)
-      return true
-    } catch (e) {
-      logger.error('Failed to checkUrl (get, auth)', { url: job.args.url, error: e });
-      lastException = e;
-    }
-  }
-
-  throw lastException;
-}
-
-const checkURLRequest = function (url, options, job) {
-  return new Promise((resolve, reject) => {
-    if (func.checkArgs(job, 'cookies')) {
-      let j = request.jar()
-      let cookie = request.cookie(job.args.cookies)
-      j.setCookie(cookie, url)
-      options.jar = j
-    } else {
-      options.jar = true
-    }
-
-    requestLoop(url, options, 2, 1000, (err, res) => {
-      if (err) {
-        return reject(err)
-      }
-
-      if (!res || !res.hasOwnProperty('statusCode')) {
-        return reject('Can\'t resolve GET request')
-      }
-
-      try {
-        if (job.args.auth && job.args.auth.type === 'netlify' && res.statusCode === 401) {
-          return resolve()
-        } else if (res.statusCode === 0 || (res.statusCode >= 400 && !([403, 404].indexOf(res.statusCode) !== -1))) {
-          return reject('Wrong status code => ' + res.statusCode + ': ' + res.statusMessage)
-        }
-
-        return resolve()
-      } catch (e) {
-        return reject('checkUrlError: ' + (e && e.hasOwnProperty('message')) ? e.message : e)
-      }
-    })
-  })
-}
-
-const requestLoop = function (url, options, attemptsLeft, retryDelay, callback, lastError = null) {
-  if (attemptsLeft <= 0) {
-    callback((lastError ?? new Error('checkUrlError')))
-  } else {
-    request(url, options, function (error, response) {
-      const recoverableErrors = ['ESOCKETTIMEDOUT', 'ETIMEDOUT', 'ECONNRESET', 'ECONNREFUSED']
-      if (error && recoverableErrors.includes(error.code)) {
-        setTimeout((function () {
-          requestLoop(url, options, --attemptsLeft, retryDelay, callback, error)
-        }), retryDelay)
-      } else {
-        callback(error, response)
-      }
-    })
-  }
-}
-
 const handleIncapsula = async (page, maxRetries = 5) => {
   for (let attempt = 0; attempt <= maxRetries; attempt++) {
     const iframeDetected = await page.$('iframe#main-iframe');
@@ -161,12 +62,12 @@ const handleIncapsula = async (page, maxRetries = 5) => {
         await page.keyboard.type('test');
         await page.keyboard.press('Tab');
         await page.evaluate(() => window.scrollBy(0, 100));
-        await new Promise(resolve => setTimeout(resolve, 1000));
 
-        const cleared = await page.waitForFunction(
-            () => !document.querySelector('iframe#main-iframe'),
-            { timeout: 10000 }
-        ).catch(() => false);
+        const cleared = await pollUntil(page, () => !document.querySelector('iframe#main-iframe'), {
+          timeoutMs: 10000,
+          intervalMs: 200,
+          label: 'Incapsula iframe removal',
+        }).catch(() => false)
 
         if (cleared) {
           logger.info('Incapsula iframe cleared. Proceeding...');
@@ -174,12 +75,17 @@ const handleIncapsula = async (page, maxRetries = 5) => {
 
         if (attempt < maxRetries) {
           logger.warn('iframe did not disappear. Retrying page reload...');
-          await new Promise(resolve => setTimeout(resolve, 2000));
-          await Promise.all([
-            page.waitForNavigation({ waitUntil: 'networkidle2', timeout: 10000 }).catch(() => {}),
-            page.reload()
-          ]);
-          await new Promise(resolve => setTimeout(resolve, 2000));
+          await page.reload();
+          try {
+            await page.waitForLoadState('networkidle', { timeout: 10000 });
+          } catch (e) {
+            logger.warn('networkidle timeout after reload; retrying with load');
+            try {
+              await page.waitForLoadState('load', { timeout: 5000 });
+            } catch (e2) {
+              logger.warn('load timeout after reload; continuing');
+            }
+          }
         } else {
           logger.error('Incapsula iframe still present after all retries.');
         }
@@ -188,64 +94,6 @@ const handleIncapsula = async (page, maxRetries = 5) => {
   }
 }
 
-const isCertificateInterstitial = async (page) => {
-  try {
-    const currentUrl = page.url();
-    if (!currentUrl || !currentUrl.startsWith('chrome-error://')) {
-      return false;
-    }
-
-    const content = await page.content().catch(() => '');
-    if (!content) {
-      return false;
-    }
-
-    const normalizedContent = content.toLowerCase();
-    return normalizedContent.includes('your connection is not private') ||
-      normalizedContent.includes('attackers might be trying to steal your information') ||
-      normalizedContent.includes('net::err_cert');
-  } catch (error) {
-    logger.warn('Failed to detect certificate interstitial', { error });
-    return false;
-  }
-};
-
-const bypassCertificateInterstitial = async (page, url, timeoutMs) => {
-  const bypassLogContext = { targetUrl: url };
-
-  if (!(await isCertificateInterstitial(page))) {
-    return false;
-  }
-
-  logger.warn('Certificate interstitial detected, attempting bypass', bypassLogContext);
-
-  try {
-    await page.keyboard.type('thisisunsafe');
-  } catch (error) {
-    logger.error('Failed to type interstitial bypass command', { error, ...bypassLogContext });
-    return false;
-  }
-
-  try {
-    await page.waitForNavigation({
-      waitUntil: ['domcontentloaded', 'load'],
-      timeout: timeoutMs,
-    }).catch(() => {});
-  } catch (error) {
-    logger.warn('Navigation wait failed after certificate bypass attempt', { error, ...bypassLogContext });
-  }
-
-  await new Promise(resolve => setTimeout(resolve, 500));
-
-  if (await isCertificateInterstitial(page)) {
-    logger.error('Certificate interstitial bypass unsuccessful', bypassLogContext);
-    return false;
-  }
-
-  logger.info('Certificate interstitial bypassed successfully', bypassLogContext);
-  return true;
-};
-
 const saveError = async (job, jobItem, errorText) => {
   let filenameKey
   let filename
@@ -253,7 +101,7 @@ const saveError = async (job, jobItem, errorText) => {
   try {
     if (errorText && (errorText.includes('SOCKETTIMEOUT') || errorText.includes('SOCKETTIMEDOUT'))) {
       errorText = 'Diffy was unable to take the screenshot.\n' +
-        'Looks like we have overloaded your server. Please try lowering number of workers for this environment under Project Settings -> Advanced -> Performance'
+          'Looks like we have overloaded your server. Please try lowering number of workers for this environment under Project Settings -> Advanced -> Performance'
     }
 
     errorText = 'Error: ' + errorText
@@ -332,145 +180,180 @@ async function disableGifAnimation(page) {
   });
 }
 
+// Safe helpers to avoid calling into a closed target.
+const ensureOpen = (page, label = 'operation') => {
+  if (!page || (typeof page.isClosed === 'function' && page.isClosed())) {
+    throw new Error(`Page closed before ${label}`)
+  }
+}
+
+const safeEval = async (page, fn, arg, label = 'evaluate') => {
+  ensureOpen(page, label)
+  return page.evaluate(fn, arg)
+}
+
+const safeAddStyleTag = async (page, opts, label = 'addStyleTag') => {
+  ensureOpen(page, label)
+  return page.addStyleTag(opts)
+}
+
+const pollUntil = async (page, predicate, {
+  timeoutMs = 5000,
+  intervalMs = 200,
+  predicateArg,
+  label = 'condition',
+} = {}) => {
+  ensureOpen(page, label)
+  const endTime = Date.now() + timeoutMs
+
+  while (Date.now() < endTime) {
+    const result = await safeEval(page, predicate, predicateArg, `${label} evaluate`)
+    if (result) {
+      return true
+    }
+
+    const delay = Math.min(intervalMs, Math.max(0, endTime - Date.now()))
+    if (delay > 0) {
+      await page.waitForTimeout(delay)
+    }
+    ensureOpen(page, `${label} wait`)
+  }
+
+  throw new Error(`Timeout waiting for ${label}`)
+}
+
 module.exports = {
 
   perform: async (browser, job, jobItem) => {
-    let data = {};
-    let page;
-    let jsConsole = [];
-    const maxPageHeightIfError = 50000;
+    // Bounded retry in case target/session closes mid-pipeline.
+    const maxAttempts = 2
+    let attempt = 0
+    let lastErr
 
-    // try {
-    //   await checkUrl(jobItem.url, jobItem)
-    //   logger.info(jobItem.id + ':' + jobItem.breakpoint + ':' + jobItem.url,'check url done')
-    // } catch (e) {
-    //   return await saveError(job, jobItem, 'CheckURL ' + ((e && e, 'message')) ? e.message : e.toString())
-    // }
+    while (attempt < maxAttempts) {
+      attempt++
 
-    try {
-      const maxPageHeight = (Object.hasOwn(job, 'attempts') && job.attempts > 0) ? (maxPageHeightIfError / job.attempts) : maxPageHeightIfError
+      let data = {};
+      let page;
+      let jsConsole = [];
+      const maxPageHeightIfError = 50000;
 
-      page = await browser.newPage();
+      try {
+        const maxPageHeight = (Object.hasOwn(job, 'attempts') && job.attempts > 0) ? (maxPageHeightIfError / job.attempts) : maxPageHeightIfError
 
-      const needIncrease = jobItem?.args?.need_increase;
-      const timeouts = {
-        default: resolveTimeoutWithNeedIncrease(needIncrease, 30000),
-        navigation: resolveTimeoutWithNeedIncrease(needIncrease, 90 * 1000),
-        goto: resolveTimeoutWithNeedIncrease(needIncrease, 120000),
-      };
+        const viewportWidth = parseInt(jobItem.breakpoint) || 800;
+        const baseViewport = { width: viewportWidth, height: 1000 };
+        const headerState = await func.buildHeaderState(jobItem);
+        const userAgentString = headerState.userAgentString;
 
-      if (Object.hasOwn(jobItem.args, 'night_mode') && jobItem.args.night_mode) {
-        // Emulate dark mode
-        await page.emulateMediaFeatures([{ name: 'prefers-color-scheme', value: 'dark' }]);
-      }
+        const hasBasicAuth = (
+            Object.hasOwn(jobItem, 'basicAuth') && jobItem.basicAuth &&
+            Object.hasOwn(jobItem.basicAuth, 'user') && jobItem.basicAuth.user &&
+            Object.hasOwn(jobItem.basicAuth, 'password') && jobItem.basicAuth.password
+        );
 
-      if (Object.hasOwn(jobItem.args, 'retina_images') && jobItem.args.retina_images) {
-        await page.setViewport({ width: parseInt(jobItem.breakpoint), height: 1000, deviceScaleFactor: 2 })
-      }
-
-      logger.debug('browser.newPage', { jobItem })
-
-      // set global timeout and disable CSP
-      await page.setBypassCSP(true)
-      await page.setDefaultNavigationTimeout(timeouts.navigation)
-      await page.setDefaultTimeout(timeouts.default)
-
-      logger.debug('setDefaultNavigationTimeout done')
-
-      page.on('console', msg => {
-        let consoleMes
-        try {
-          consoleMes = {
-            type: msg.type(),
-            text: msg.text(),
-            location: msg.location(),
-          }
-        } catch (e) {
-          consoleMes = {
-            type: e.type(),
-            text: e.text(),
-            location: e.location(),
-          }
-        }
-
-        jsConsole.push(consoleMes)
-      })
-
-      // Remove all browser Cookies.
-      const client = await page.target().createCDPSession();
-      await client.send('Network.clearBrowserCookies');
-      await new Promise(resolve => setTimeout(resolve, 1000));
-
-      await func.setHeaders(page, jobItem)
-      logger.debug('setHeaders done')
-
-      if (!Object.hasOwn(jobItem, 'url') || !Object.hasOwn(jobItem, 'breakpoint')) {
-        throw new Error('Cannot find url or breakpoint options')
-      }
-
-      let url = jobItem.url;
-
-      if (jobItem.url && jobItem.base_url) {
-        // Base URL can have GET parameters. We need to merge them with url.
-        let pageUrl = new URL(jobItem.url);
-        let pageUrlParameters = pageUrl.searchParams;
-        let pageUrlHash = pageUrl.hash;
-
-        let baseUrl = new URL(jobItem.base_url);
-        let baseUrlParameters = baseUrl.searchParams;
-
-        // We override base URL parameters with ones from the page.
-        pageUrlParameters.forEach((value, key) => {
-          baseUrlParameters.set(key, value);
+        page = await browser.newPage({
+          viewport: baseViewport,
+          bypassCSP: true,
+          ignoreHTTPSErrors: true,
+          userAgent: userAgentString,
+          deviceScaleFactor: (Object.hasOwn(jobItem.args, 'retina_images') && jobItem.args.retina_images) ? 2 : 1,
+          locale: 'en-US',
+          timezoneId: 'UTC',
+          hasTouch: false,
+          ...(hasBasicAuth ? {
+            httpCredentials: {
+              username: jobItem.basicAuth.user,
+              password: jobItem.basicAuth.password,
+            }
+          } : {}),
         });
 
-        url = jobItem.url.replace(/[\?#].*$/, '');
+        const browserContext = page.context();
+        await func.setHeaders(page, jobItem, headerState);
 
-        const parametersString = baseUrlParameters.toString();
-        if (parametersString) {
-          url += '?' + parametersString;
+        if (Object.hasOwn(jobItem.args, 'night_mode') && jobItem.args.night_mode) {
+          await page.emulateMedia({colorScheme: 'dark'});
         }
 
-        if (pageUrlHash) {
-          url += pageUrlHash;
+        logger.debug('browser.newPage', {jobItem})
+
+        await page.setDefaultNavigationTimeout(90000)
+        await page.setDefaultTimeout(30000)
+
+        logger.debug('setDefaultNavigationTimeout done')
+
+        page.on('console', msg => {
+          let consoleMes
+          try {
+            consoleMes = {
+              type: msg.type(),
+              text: msg.text(),
+              location: msg.location(),
+            }
+          } catch (e) {
+            consoleMes = {
+              type: e.type(),
+              text: e.text(),
+              location: e.location(),
+            }
+          }
+
+          jsConsole.push(consoleMes)
+        })
+
+        await browserContext.clearCookies();
+        logger.debug('setHeaders prepared', {
+          userAgent: userAgentString,
+          extraHeaders: headerState?.headers || {}
+        })
+
+        if (!Object.hasOwn(jobItem, 'url') || !Object.hasOwn(jobItem, 'breakpoint')) {
+          throw new Error('Cannot find url or breakpoint options')
         }
-      }
 
-      if (
-        Object.hasOwn(jobItem, 'basicAuth') && jobItem.basicAuth &&
-        Object.hasOwn(jobItem.basicAuth, 'user') && jobItem.basicAuth.user &&
-        Object.hasOwn(jobItem.basicAuth, 'password') && jobItem.basicAuth.password
-      ) {
-        await page.authenticate({ username: jobItem.basicAuth.user, password: jobItem.basicAuth.password })
+        let url = jobItem.url;
 
-        if (url.startsWith('http://')) {
-          const base64 = Buffer.from(`${jobItem.basicAuth.user}:${jobItem.basicAuth.password}`).toString('base64');
-          const targetHost = new URL(jobItem.base_url).host;
+        if (jobItem.url && jobItem.base_url) {
+          // Base URL can have GET parameters. We need to merge them with url.
+          let pageUrl = new URL(jobItem.url);
+          let pageUrlParameters = pageUrl.searchParams;
+          let pageUrlHash = pageUrl.hash;
 
-          await page.setRequestInterception(true);
+          let baseUrl = new URL(jobItem.base_url);
+          let baseUrlParameters = baseUrl.searchParams;
 
-          page.on('request', (request) => {
-            const headers = {
-              ...request.headers(),
-              Authorization: `Basic ${base64}`,
-            };
+          // We override base URL parameters with ones from the page.
+          pageUrlParameters.forEach((value, key) => {
+            baseUrlParameters.set(key, value);
+          });
 
-            let reqUrl = request.url();
-            const currentHost = (() => {
+          url = jobItem.url.replace(/[\?#].*$/, '');
+
+          const parametersString = baseUrlParameters.toString();
+          if (parametersString) {
+            url += '?' + parametersString;
+          }
+
+          if (pageUrlHash) {
+            url += pageUrlHash;
+          }
+        }
+
+        const callRailBlockEnabled = Object.hasOwn(jobItem, 'project_id') && jobItem.project_id === 21791;
+
+        let basicAuthRouteConfig = null;
+        if (hasBasicAuth && url.startsWith('http://')) {
+          basicAuthRouteConfig = {
+            header: `Basic ${Buffer.from(`${jobItem.basicAuth.user}:${jobItem.basicAuth.password}`).toString('base64')}`,
+            targetHost: (() => {
               try {
-                return new URL(reqUrl).host;
-              } catch {
+                return new URL(jobItem.base_url).host;
+              } catch (e) {
                 return null;
               }
-            })();
-
-            // Only downgrade to HTTP if host matches base_url
-            if (currentHost === targetHost) {
-              reqUrl = reqUrl.replace(/^https:/, 'http:');
-            }
-
-            request.continue({ headers, url: reqUrl });
-          });
+            })()
+          };
 
           page.on('response', async (res) => {
             const status = res.status();
@@ -486,71 +369,214 @@ module.exports = {
             }
           });
         }
-      }
 
-      // Add new cookies.
-      let cookies = await func.addCookies(jobItem)
-      logger.debug('addCookies done')
+        const deterrenceState = headerState?.deterrenceBypass
+        const selectiveHeaderRoutingEnabled = callRailBlockEnabled || basicAuthRouteConfig || deterrenceState?.enabled
 
-      const authCookies = await func.auth(page, jobItem).catch((err) => {
-        data.auth_error = err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err
-      })
+        if (selectiveHeaderRoutingEnabled) {
+          await page.route('**/*', async (route) => {
+            const request = route.request();
+            const requestUrl = request.url();
 
-      logger.debug('auth done')
+            if (callRailBlockEnabled) {
+              try {
+                if (/swap_session\.json/i.test(new URL(requestUrl).pathname)) {
+                  route.abort().catch((error) => {
+                    logger.warn('Failed to abort blocked request', { error, requestUrl });
+                  });
+                  return;
+                }
+              } catch (_) {}
+            }
 
-      if (authCookies) {
-        logger.debug('authCookies', { authCookies })
-        cookies = cookies.concat(authCookies)
-      }
+            let headersOverride = null;
+            let overriddenUrl = requestUrl;
 
-      if (cookies) {
-        await page.setCookie(...cookies)
-      }
+            const ensureHeadersCopy = () => {
+              if (!headersOverride) {
+                headersOverride = { ...request.headers() };
+              }
+              return headersOverride;
+            };
 
-      if (Object.hasOwn(jobItem, 'project_id') && jobItem.project_id === 21791) {
-        // @see https://support.callrail.com/hc/en-us/articles/5711492051085-Preventing-a-number-from-swapping-on-a-website
-        logger.debug('Apply calltrkNoswap')
+            if (basicAuthRouteConfig) {
+              const headers = ensureHeadersCopy();
+              headers.Authorization = basicAuthRouteConfig.header;
 
-        await page.setRequestInterception(true);
-        page.on('request', interceptedRequest => {
-          if (interceptedRequest.url().endsWith('swap_session.json')) {
-            interceptedRequest.abort();
-          } else {
-            interceptedRequest.continue();
-          }
-        });
-      }
+              try {
+                const host = new URL(requestUrl).host;
+                if (host && basicAuthRouteConfig.targetHost && host === basicAuthRouteConfig.targetHost) {
+                  overriddenUrl = overriddenUrl.replace(/^https:/, 'http:');
+                }
+              } catch (_) {}
+            }
 
-      let response;
+            if (deterrenceState && func.shouldApplyDeterrenceHeader(deterrenceState, requestUrl, request.resourceType())) {
+              const headers = ensureHeadersCopy();
+              headers[deterrenceState.headerName] = deterrenceState.headerValue;
+            }
 
-      try {
-        response = await page.goto(url, { timeout: timeouts.goto, waitUntil: ['networkidle2'] })
+            const hasHeaderOverrides = headersOverride && Object.keys(headersOverride).length > 0;
+            const needsUrlOverride = overriddenUrl !== requestUrl;
 
-        await handleIncapsula(page);
-      } catch (err) {
-        logger.warn('page was not loaded by networkidle2', { url })
+            const continueWithLogging = async (options) => {
+              try {
+                await route.continue(options);
+              } catch (error) {
+                logger.warn('Failed to continue request', {
+                  requestUrl,
+                  requestMethod: request.method(),
+                  resourceType: request.resourceType(),
+                  isNavigationRequest: typeof request.isNavigationRequest === 'function'
+                    ? request.isNavigationRequest()
+                    : undefined,
+                  overrides: {
+                    headerKeys: headersOverride ? Object.keys(headersOverride) : [],
+                    urlChanged: Boolean(options && options.url && options.url !== requestUrl),
+                  },
+                  error: formatLoggedError(error)
+                });
+              }
+            };
+
+            const headerOverrideOptions = hasHeaderOverrides ? { headers: headersOverride } : undefined;
+
+            if (!needsUrlOverride) {
+              await continueWithLogging(headerOverrideOptions);
+              return;
+            }
+
+            let originalProtocol;
+            let overrideProtocol;
+            try {
+              originalProtocol = new URL(requestUrl).protocol;
+              overrideProtocol = new URL(overriddenUrl).protocol;
+            } catch (error) {
+              logger.warn('Failed to parse URLs for override; continuing without URL change', {
+                requestUrl,
+                attemptedUrl: overriddenUrl,
+                error: formatLoggedError(error)
+              });
+              await continueWithLogging(headerOverrideOptions);
+              return;
+            }
+
+            if (originalProtocol === overrideProtocol) {
+              await continueWithLogging({
+                ...headerOverrideOptions,
+                url: overriddenUrl
+              });
+              return;
+            }
+
+            const fetchHeaders = hasHeaderOverrides ? headersOverride : request.headers();
+            const requestMethod = request.method();
+            const postData = request.postData();
+
+            const fetchOptions = {
+              url: overriddenUrl,
+              headers: fetchHeaders,
+              ...(requestMethod && requestMethod !== 'GET' ? { method: requestMethod } : {}),
+              ...(postData ? { postData } : {})
+            };
+
+            let fetchResponse;
+            try {
+              fetchResponse = await route.fetch(fetchOptions);
+            } catch (error) {
+              logger.warn('Failed to fetch overridden protocol request; falling back to original URL', {
+                requestUrl,
+                attemptedUrl: overriddenUrl,
+                requestMethod,
+                resourceType: request.resourceType(),
+                error: formatLoggedError(error)
+              });
+              await continueWithLogging(headerOverrideOptions);
+              return;
+            }
+
+            try {
+              await route.fulfill({ response: fetchResponse });
+            } catch (error) {
+              logger.warn('Failed to fulfill fetched response; continuing without URL override', {
+                requestUrl,
+                attemptedUrl: overriddenUrl,
+                requestMethod,
+                resourceType: request.resourceType(),
+                error: formatLoggedError(error)
+              });
+              await continueWithLogging(headerOverrideOptions);
+            }
+          });
+        }
+
+        // Add new cookies.
+        let cookies = await func.addCookies(jobItem)
+        logger.debug('addCookies done')
+
+        const authCookies = await func.auth(page, jobItem).catch((err) => {
+          const message = (err && Object.hasOwn(err, 'message')) ? err.message : err;
+          data.auth_error = `${err?.name || 'AuthError'}: ${message}`;
+          return null;
+        })
+
+        logger.debug('auth done')
+
+        if (authCookies) {
+          logger.debug('authCookies', {authCookies})
+          cookies = cookies.concat(authCookies)
+        }
+
+        if (cookies?.length) {
+          await browserContext.addCookies(cookies)
+        }
+
+        let response;
 
         try {
-          response = await page.goto(url, { timeout: timeouts.goto, waitUntil: ['load', 'domcontentloaded'] })
+          await page.waitForTimeout(func.random(120, 380));
+          response = await page.goto(url, {waitUntil: 'networkidle'})
+
+          await handleIncapsula(page);
+          await func.handleCloudflareChallenge(page, {frameWaitMs: 8000, retryDelayMs: 2500}).catch((error) => {
+            logger.warn('Cloudflare challenge handling failed', {error})
+          })
+          const unresolvedChallenge = await page.evaluate(() => {
+            const bodyText = document.body?.innerText || ''
+            return bodyText.includes('Please unblock challenges.cloudflare.com')
+          }).catch(() => false)
+          if (unresolvedChallenge) {
+            throw new Error('Cloudflare challenge unresolved: Please unblock challenges.cloudflare.com')
+          }
         } catch (err) {
-          logger.error('page was not loaded by load or domcontentloaded', { error: err, url })
+          logger.debug('page was not loaded by networkidle')
+
+          try {
+            response = await page.goto(url, {waitUntil: 'load'})
+            await page.waitForLoadState('domcontentloaded', {timeout: 120000}).catch(() => {
+            })
+            await func.handleCloudflareChallenge(page, {frameWaitMs: 8000, retryDelayMs: 2500}).catch((error) => {
+              logger.warn('Cloudflare challenge handling failed (retry branch)', {error})
+            })
+            const unresolvedChallenge = await page.evaluate(() => {
+              const bodyText = document.body?.innerText || ''
+              return bodyText.includes('Please unblock challenges.cloudflare.com')
+            }).catch(() => false)
+            if (unresolvedChallenge) {
+              throw new Error('Cloudflare challenge unresolved after reload')
+            }
+          } catch (err) {
+            logger.warn('page was not loaded by load or domcontentloaded', {error: err, url})
+          }
         }
-      }
 
-      const certificateBypassed = await bypassCertificateInterstitial(page, url, timeouts.goto);
-      if (certificateBypassed) {
-        await new Promise(resolve => setTimeout(resolve, 1000));
-      } else if (await isCertificateInterstitial(page)) {
-        throw new Error('Unable to bypass certificate interstitial');
-      }
+        logger.debug('page loaded done')
 
-      logger.debug('page loaded done')
+        // Disable animation / transition (exclude diff from animation)
+        logger.debug('disable css animation')
 
-      // Disable animation / transition (exclude diff from animation)
-      logger.debug('disable css animation')
-
-      await page.addStyleTag({
-        content: `
+        await safeAddStyleTag(page, {
+          content: `
             *, *::after, *::before {
               transition-delay: 0s !important;
               transition-duration: 0s !important;
@@ -561,276 +587,320 @@ module.exports = {
               color-adjust: exact !important;
             }
           `
-      }).catch((e) => logger.warn('Failed to add style tag to disable animation', { error: e }))
+        }).catch((e) => logger.warn('Failed to add style tag to disable animation', {error: e}))
 
-      try {
-        await disableGifAnimation(page)
-      } catch (e) {
-        logger.warn('Failed to disable GIF animation', { error: e })
-      }
-
-      await page.setViewport({ width: parseInt(jobItem.breakpoint), height: 1000 })
-      await new Promise(resolve => setTimeout(resolve, 1000));
-      logger.debug('page.goto done')
-
-      await page.evaluate(() => document.fonts.ready)
-      await page.waitForFunction(() => document.readyState === 'complete');
-      await new Promise(resolve => setTimeout(resolve, 1000));
-
-      // @see https://github.com/ygerasimov/diffy-pm/issues/250 (wp-rocket fix)
-      await page.evaluate(() => {
         try {
-          window.dispatchEvent(new Event('touchstart'));
-          window.document.dispatchEvent(new Event('touchstart'));
-        } catch (e) {}
-      });
+          ensureOpen(page, 'disableGifAnimation');
+          await disableGifAnimation(page)
+        } catch (e) {
+          logger.warn('Failed to disable GIF animation', {error: e})
+        }
 
-      await func.addCssCode(page, jobItem)
-      logger.debug('addCssCode done')
+        if (!page.isClosed()) {
+          await page.setViewportSize({width: parseInt(jobItem.breakpoint), height: 1000})
+          await page.waitForTimeout(1000)
+        }
+        logger.debug('page.goto done')
 
-      // #see https://github.com/ygerasimov/diffy-pm/issues/339
-      if (Object.hasOwn(jobItem, 'project_id') && jobItem.project_id === 20882) {
-        await func.cutElements(page, jobItem)
-      }
+        await safeEval(page, () => {
+          if (!document.fonts || typeof document.fonts.ready?.then !== 'function') {
+            return null
+          }
+          return document.fonts.ready
+        }, undefined, 'fonts.ready wait')
+        await pollUntil(page, () => document.readyState === 'complete', {
+          timeoutMs: 30000,
+          intervalMs: 250,
+          label: 'document.readyState === complete',
+        }).catch((error) => {
+          logger.warn('document.readyState wait failed', {error});
+        })
+        await page.waitForTimeout(1000)
 
-      await func.autoScroll(page, jobItem)
-      logger.debug('autoScroll done')
+        // @see https://github.com/ygerasimov/diffy-pm/issues/250 (wp-rocket fix)
+        await safeEval(page, () => {
+          try {
+            window.dispatchEvent(new Event('touchstart'));
+            window.document.dispatchEvent(new Event('touchstart'));
+          } catch (e) {
+          }
+        }, undefined, 'wp-rocket-fix');
 
-      if (Object.hasOwn(jobItem.args, 'stabilization') && jobItem.args.stabilization) {
-          await (async () => {
+        await func.addCssCode(page, jobItem)
+        logger.debug('addCssCode done')
+
+        // #see https://github.com/ygerasimov/diffy-pm/issues/339
+        if (Object.hasOwn(jobItem, 'project_id') && jobItem.project_id === 20882) {
+          await func.cutElements(page, jobItem)
+        }
+
+        await func.autoScroll(page, jobItem)
+        logger.debug('autoScroll done')
+
+        const stabilizationEnabled = Boolean(Object.hasOwn(jobItem.args, 'stabilization') && jobItem.args.stabilization)
+
+        if (stabilizationEnabled && Object.hasOwn(jobItem.args, 'stabilization_code') && jobItem.args.stabilization_code) {
+          try {
+            await (async () => {
               await eval(jobItem.args.stabilization_code);
-          })();
-      }
+            })();
+          } catch (error) {
+            logger.warn('stabilization_code execution failed', { error: error?.message || String(error) })
+          }
+        }
 
-      let page_height = await func.updatePageViewport(page, jobItem, maxPageHeight)
-      logger.debug('updatePageViewport done', { page_height })
+        const initialViewportHeight = await func.updatePageViewport(page, jobItem, maxPageHeight)
+        logger.debug('updatePageViewport done', {page_height: initialViewportHeight})
 
-      if (Object.hasOwn(jobItem.args, 'stabilization') && jobItem.args.stabilization) {
-        await page.evaluate(async () => {
+        if (stabilizationEnabled) {
+          await page.evaluate(async () => {
 
-          const stabilizeHeight = async (elementsHeights, level) => {
-            for (const element of elementsHeights) {
-              if (document.body.contains(element.node)) {
-                if (
-                    element.height !== element.node.offsetHeight &&
-                    element.viewportRatio >= 0.40
-                ) {
-                  element.node.style.height = element.height + 'px'
-                  element.node.style.maxHeight = element.height + 'px'
-                  element.node.style.minHeight = element.height + 'px'
+            const stabilizeHeight = async (elementsHeights, level) => {
+              for (const element of elementsHeights) {
+                if (document.body.contains(element.node)) {
+                  if (
+                      element.height !== element.node.offsetHeight &&
+                      element.viewportRatio >= 0.40
+                  ) {
+                    element.node.style.height = element.height + 'px'
+                    element.node.style.maxHeight = element.height + 'px'
+                    element.node.style.minHeight = element.height + 'px'
 
-                  if (element.node.scrollHeight === element.node.offsetHeight) {
-                    continue
+                    if (element.node.scrollHeight === element.node.offsetHeight) {
+                      continue
+                    }
                   }
-                }
 
-                if (element.childNodes.length) {
-                  await stabilizeHeight(element.childNodes, level + 1)
+                  if (element.childNodes.length) {
+                    await stabilizeHeight(element.childNodes, level + 1)
+                  }
                 }
               }
             }
-          }
 
-          await stabilizeHeight(window.diffyElementsHeights ?? [], 1);
+            await stabilizeHeight(window.diffyElementsHeights ?? [], 1);
+          })
+        }
+
+        if (stabilizationEnabled) {
+          await func.hideBanners(page, { args: { elements: ['iframe[src*="google.com/maps"]'] } })
+        }
+
+        await func.delayBeforeScreenshot(page, jobItem)
+
+        await func.addJsCode(page, jobItem)
+        logger.debug('addJsCode done')
+
+        logger.debug('delayBeforeScreenshot done')
+        ensureOpen(page, 'cutElements')
+        const is_cut = await func.cutElements(page, jobItem)
+        if (is_cut) {
+          // We need decrease height after cut.
+          if (!page.isClosed()) {
+            await page.setViewportSize({width: parseInt(jobItem.breakpoint), height: 100})
+            await page.waitForTimeout(1000)
+            await func.updatePageViewport(page, jobItem, maxPageHeight)
+          }
+        }
+        logger.debug('cutElements done')
+
+        await func.addFixtures(page, jobItem)
+        logger.debug('addFixtures done')
+
+        await func.hideBanners(page, jobItem)
+        logger.debug('hideBanners done')
+
+        // Recalculate page height after modifications.
+        if (!page.isClosed()) {
+          await page.setViewportSize({width: parseInt(jobItem.breakpoint), height: 100})
+          await page.waitForTimeout(1000)
+          await func.updatePageViewport(page, jobItem, maxPageHeight)
+        }
+
+        await func.autoScroll(page, jobItem)
+        logger.debug('double autoScroll done')
+
+        const pageHeight = await func.updatePageViewport(page, jobItem, maxPageHeight)
+
+        data.pageArea = pageHeight * jobItem.breakpoint
+
+        logger.debug('updatePageViewport done')
+
+        const is_crop = await func.cropElement(page, jobItem)
+
+        logger.debug('cropElement done')
+
+        const filenameKey = Math.floor(Date.now() / 1000) + '-' + (func.random(0, 999999999)).toString()
+        let filename = '/tmp/screenshot-' + filenameKey + '.png'
+
+        const htmlFilename = '/tmp/html-' + filenameKey + '.html'
+
+        let mhtmlFilename = '';
+        if (Object.hasOwn(jobItem, 'mhtml') && jobItem.mhtml) {
+          mhtmlFilename = '/tmp/mhtml-' + filenameKey + '.mhtml'
+        }
+
+        const jsConsoleFilename = '/tmp/jsConsole-' + filenameKey + '.json'
+        let thumbnailFilepath = filename.replace('.png', '-thumbnail.png')
+
+        logger.debug('start screenshot')
+
+        await page.screenshot({
+          path: filename,
+          captureBeyondViewport: false,
         })
 
-        // hide google maps
-        await func.hideBanners(page, { args: { elements: ['iframe[src*="google.com/maps"]'] } })
-      }
+        logger.debug('screenshot done')
+        const pageHtml = await func.getPageHtml(page)
+        logger.debug('pageHtml done')
 
-      await func.delayBeforeScreenshot(page, jobItem)
-
-      await func.addJsCode(page, jobItem)
-      logger.debug('addJsCode done')
-
-      logger.debug('delayBeforeScreenshot done')
-      const is_cut = await func.cutElements(page, jobItem)
-      if (is_cut) {
-        // We need decrease height after cut.
-        await page.setViewport({ width: parseInt(jobItem.breakpoint), height: 100 })
-        await func.updatePageViewport(page, jobItem, maxPageHeight)
-      }
-      logger.debug('cutElements done')
-
-      await func.addFixtures(page, jobItem)
-      logger.debug('addFixtures done')
-
-      await func.hideBanners(page, jobItem)
-      logger.debug('hideBanners done')
-
-      // Recalculate page height after modifications.
-      await page.setViewport({ width: parseInt(jobItem.breakpoint), height: 100 })
-      await func.updatePageViewport(page, jobItem, maxPageHeight)
-
-      await func.autoScroll(page, jobItem)
-      logger.debug('double autoScroll done')
-      const pageHeight = await func.updatePageViewport(page, jobItem, maxPageHeight)
-
-      data.pageArea = pageHeight * jobItem.breakpoint
-
-      logger.debug('updatePageViewport done')
-
-      const is_crop = await func.cropElement(page, jobItem)
-
-      logger.debug('cropElement done')
-
-      const filenameKey = Math.floor(Date.now() / 1000) + '-' + (func.random(0, 999999999)).toString()
-      let filename = '/tmp/screenshot-' + filenameKey + '.png'
-
-      const htmlFilename = '/tmp/html-' + filenameKey + '.html'
-
-      let mhtmlFilename = '';
-      if (Object.hasOwn(jobItem, 'mhtml') && jobItem.mhtml) {
-        mhtmlFilename = '/tmp/mhtml-' + filenameKey + '.mhtml'
-      }
-
-      const jsConsoleFilename = '/tmp/jsConsole-' + filenameKey + '.json'
-      let thumbnailFilepath = filename.replace('.png', '-thumbnail.png')
-
-      logger.debug('start screenshot')
-
-      await page.screenshot({
-        path: filename,
-        captureBeyondViewport: false,
-      })
-
-      logger.debug('screenshot done')
-      const pageHtml = await func.getPageHtml(page)
-      logger.debug('pageHtml done')
-
-      let pageMhtml = ''
-      if (mhtmlFilename) {
-        pageMhtml = await func.getPageMhtml(page)
-        logger.debug('pageMhtml done', { jobItem })
-      }
-
-      if (is_crop) {
-        await thumbnail.crop(filename, is_crop)
-        data.pageArea = is_crop.height * is_crop.width
-      }
-
-      await page.close()
-      logger.debug('page close done')
-
-      // check webp format
-      const screenshotSize = await func.getImageSize(filename)
-      let webpWasUsed = false
-
-      if (screenshotSize.height < 16000 && screenshotSize.width < 16000) {
-        const filenameWebp = filename.replace('.png', '.webp')
-
-        await thumbnail.webp(filename, filenameWebp)
-
-        filename = filenameWebp
-        thumbnailFilepath = thumbnailFilepath.replace('.png', '.webp')
-
-        webpWasUsed = true
-      }
-
-      logger.debug('screenshot created')
-
-      if (jobItem.local) {
-        const fs = require('node:fs');
-        try {
-          fs.writeFileSync(htmlFilename, pageHtml);
-        } catch (err) {
-          logger.error('Failed to write file', { error: err });
+        let pageMhtml = ''
+        if (mhtmlFilename) {
+          pageMhtml = await func.getPageMhtml(page)
+          logger.debug('pageMhtml done', {jobItem})
         }
+
+        if (is_crop) {
+          await thumbnail.crop(filename, is_crop)
+          data.pageArea = is_crop.height * is_crop.width
+        }
+
+        if (page && !page.isClosed()) {
+          await page.close()
+        }
+        logger.debug('page close done')
+        page = null
+
+        // check webp format
+        const screenshotSize = await func.getImageSize(filename)
+        let webpWasUsed = false
+
+        if (screenshotSize.height < 16000 && screenshotSize.width < 16000) {
+          const filenameWebp = filename.replace('.png', '.webp')
+
+          await thumbnail.webp(filename, filenameWebp)
+
+          filename = filenameWebp
+          thumbnailFilepath = thumbnailFilepath.replace('.png', '.webp')
+
+          webpWasUsed = true
+        }
+
+        logger.debug('screenshot created')
+
+        if (jobItem.local) {
+          const fs = require('node:fs');
+          try {
+            fs.writeFileSync(htmlFilename, pageHtml);
+          } catch (err) {
+            logger.error('Failed to write file', {error: err});
+          }
+
+          if (mhtmlFilename) {
+            try {
+              fs.writeFileSync(mhtmlFilename, pageMhtml);
+            } catch (err) {
+              logger.error('Failed to write MHTML file', {error: err});
+            }
+          }
+
+          try {
+            fs.writeFileSync(jsConsoleFilename, JSON.stringify(jsConsole));
+          } catch (err) {
+            logger.error('Failed to write file', {error: err});
+          }
+
+          return {
+            screenshot: filename,
+            html: htmlFilename,
+            mhtml: mhtmlFilename,
+            jsConsole: jsConsoleFilename
+          }
+        }
+
+        const s3Url = await uploadS3.upload(filename).catch((err) => {
+          logger.error('Failed to upload file to S3', {error: err});
+          throw new Error('Can\'t upload screenshot: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
+        })
+
+        logger.debug('uploadS3 done')
+
+        await thumbnail.generateImageThumbnail(filename, thumbnailFilepath).catch((err) => {
+          throw new Error('Can\'t generate thumbnail: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
+        })
+
+        logger.debug('generateImageThumbnail done')
+
+        const s3UrlThumbnail = await uploadS3.upload(thumbnailFilepath).catch((err) => {
+          throw new Error('Can\'t upload thumbnail: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
+        })
+
+        logger.debug('uploadS3Thumbnail done')
+
+        const s3HtmlUrl = await uploadS3.uploadFileString(htmlFilename, pageHtml).catch((err) => {
+          throw new Error('Can\'t upload html file: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
+        })
+
+        logger.debug('uploadHtmlFileString done')
+
+        let s3MhtmlUrl = ''
 
         if (mhtmlFilename) {
-          try {
-            fs.writeFileSync(mhtmlFilename, pageMhtml);
-          } catch (err) {
-            logger.error('Failed to write MHTML file', { error: err });
-          }
+          s3MhtmlUrl = await uploadS3.uploadFileString(mhtmlFilename, pageMhtml).catch((err) => {
+            throw new Error('Can\'t upload mhtml file: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
+          })
+
+          logger.debug('uploadMhtmlFileString done', {job_item: jobItem})
         }
 
-        try {
-          fs.writeFileSync(jsConsoleFilename, JSON.stringify(jsConsole));
-        } catch (err) {
-          logger.error('Failed to write file', { error: err });
-        }
-
-        return {
-          screenshot: filename,
-          html: htmlFilename,
-          mhtml: mhtmlFilename,
-          jsConsole: jsConsoleFilename
-        }
-      }
-
-      const s3Url = await uploadS3.upload(filename).catch((err) => {
-        logger.error('Failed to upload file to S3', { error: err });
-        throw new Error('Can\'t upload screenshot: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
-      })
-
-      logger.debug('uploadS3 done')
-
-      await thumbnail.generateImageThumbnail(filename, thumbnailFilepath).catch((err) => {
-        throw new Error('Can\'t generate thumbnail: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
-      })
-
-      logger.debug('generateImageThumbnail done')
-
-      const s3UrlThumbnail = await uploadS3.upload(thumbnailFilepath).catch((err) => {
-        throw new Error('Can\'t upload thumbnail: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
-      })
-
-      logger.debug('uploadS3Thumbnail done')
-
-      const s3HtmlUrl = await uploadS3.uploadFileString(htmlFilename, pageHtml).catch((err) => {
-        throw new Error('Can\'t upload html file: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
-      })
-
-      logger.debug('uploadHtmlFileString done')
-
-      let s3MhtmlUrl = ''
-
-      if (mhtmlFilename) {
-        s3MhtmlUrl = await uploadS3.uploadFileString(mhtmlFilename, pageMhtml).catch((err) => {
-          throw new Error('Can\'t upload mhtml file: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
+        const s3JsConsoleUrl = await uploadS3.uploadFileString(jsConsoleFilename, JSON.stringify(jsConsole)).catch((err) => {
+          throw new Error('Can\'t upload jsConsole file: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
         })
 
-        logger.debug('uploadMhtmlFileString done', { job_item: jobItem })
-      }
+        logger.debug('uploadJsConsoleFileString done')
 
-      const s3JsConsoleUrl = await uploadS3.uploadFileString(jsConsoleFilename, JSON.stringify(jsConsole)).catch((err) => {
-        throw new Error('Can\'t upload jsConsole file: ' + err.name + ': ' + (err && Object.hasOwn(err, 'message')) ? err.message : err)
-      })
+        // Not need to remove "htmlFilename/mhtmlFilename" because we use stream and not creating real file.
+        // Async remove files.
+        await func.removeFile(filename)
+        await func.removeFile(thumbnailFilepath)
 
-      logger.debug('uploadJsConsoleFileString done')
-
-      // Not need to remove "htmlFilename/mhtmlFilename" because we use stream and not creating real file.
-      // Async remove files.
-      await func.removeFile(filename)
-      await func.removeFile(thumbnailFilepath)
-
-      if (webpWasUsed) {
-        await func.removeFile(filename.replace('.webp', '.png'))
-      }
-
-      return sendResult(job, jobItem, {
-        'full': s3Url,
-        'thumbnail': s3UrlThumbnail,
-        'html': s3HtmlUrl,
-        'mhtml': s3MhtmlUrl,
-        'jsConsole': s3JsConsoleUrl,
-        'data': data,
-        'log_data': '',
-        'status': response ? response.status() : null,
-      })
-    } catch (err) {
-      logger.error('perform error:', { error: err })
-
-      if (page !== null) {
-        try {
-          await page.close()
-        } catch (e) {
-          logger.error('Failed to close page', { error: e })
+        if (webpWasUsed) {
+          await func.removeFile(filename.replace('.webp', '.png'))
         }
-      }
 
-      return sendError(job, (err && Object.hasOwn(err, 'message')) ? err.message : err.toString(), jobItem)
+        return sendResult(job, jobItem, {
+          'full': s3Url,
+          'thumbnail': s3UrlThumbnail,
+          'html': s3HtmlUrl,
+          'mhtml': s3MhtmlUrl,
+          'jsConsole': s3JsConsoleUrl,
+          'data': data,
+          'log_data': '',
+          'status': response ? response.status() : null,
+        })
+      } catch (err) {
+        logger.error('perform error:', {error: err})
+
+        if (page) {
+          try {
+            await page.close()
+          } catch (e) {
+            logger.error('Failed to close page', {error: e})
+          }
+          page = null
+        }
+
+        // Retry once for transient target/session closed errors
+        const msg = (err && Object.hasOwn(err, 'message')) ? err.message : err.toString()
+        lastErr = msg
+        const transient = /Target closed|Session closed|Protocol error/.test(msg)
+        if (attempt < maxAttempts && transient) {
+          logger.warn('Retrying after transient closure', {attempt, msg})
+          continue
+        }
+        return sendError(job, msg, jobItem)
+      }
+      return sendError(job, lastErr || 'Unknown error', jobItem)
     }
   },
