+++ lib/thumbnail.js
@@ -1,6 +1,38 @@
 const sharp = require('sharp');
 const thumbnailWidth = 220;
 
+const escapeSvgText = (text = '') => {
+    return text
+        .replace(/&/g, '&amp;')
+        .replace(/</g, '&lt;')
+        .replace(/>/g, '&gt;')
+        .replace(/"/g, '&quot;')
+        .replace(/'/g, '&#39;');
+};
+
+const wrapText = (text = '', maxLineLength = 90) => {
+    if (!text.length) {
+        return [''];
+    }
+
+    const result = [];
+    text.split(/\r?\n/).forEach(line => {
+        if (!line.length) {
+            result.push('');
+            return;
+        }
+
+        let remaining = line;
+        while (remaining.length > maxLineLength) {
+            result.push(remaining.slice(0, maxLineLength));
+            remaining = remaining.slice(maxLineLength);
+        }
+        result.push(remaining);
+    });
+
+    return result;
+};
+
 module.exports = {
     generateImageThumbnail: async function (file, resultFile) {
         return sharp(file)
@@ -21,8 +53,8 @@ module.exports = {
     },
 
     createErrorImage: async function (filename, errorText, width = 1024) {
-        let textArr = errorText.replace(/'/gi, "`").match(/.{1,90}/g);
-        errorText = textArr.join("\n");
+        const sanitizedLines = wrapText((errorText || '').toString());
+        const escapedText = sanitizedLines.map(line => escapeSvgText(line)).join("\n");
 
         return sharp({
             create: {
@@ -33,7 +65,7 @@ module.exports = {
             }
         })
             .composite([{
-                input: Buffer.from(`<svg width="${width}" height="600"><text x="10" y="20" font-family="Verdana" font-size="20" fill="black">${errorText}</text></svg>`),
+                input: Buffer.from(`<svg width="${width}" height="600"><text x="10" y="20" font-family="Verdana" font-size="20" fill="black">${escapedText}</text></svg>`),
                 top: 0,
                 left: 0
             }])
