+++ package.json
@@ -1,5 +1,5 @@
 {
-    "name": "Diffy local screenshot Chrome 111",
+    "name": "diffy-playwright-worker",
     "version": "1.0.0",
     "description": "Diffy local screenshots",
     "bin": {
@@ -15,16 +15,16 @@
         "@aws-sdk/client-sqs": "^3.637.0",
         "@aws-sdk/credential-providers": "^3.637.0",
         "bluebird": "^3.7.2",
+        "dotenv": "^17.2.3",
         "fs-extra": "^8.1.0",
         "iltorb": "^2.4.4",
         "minimist": "^1.2.8",
+        "playwright": "^1.56.0",
+        "playwright-core": "^1.49.0",
+        "playwright-extra": "^4.3.6",
+        "playwright-extra-plugin-stealth": "^0.0.1",
         "process": "^0.11.10",
         "proxy-chain": "^2.4.0",
-        "puppeteer": "^23.9.0",
-        "puppeteer-core": "^23.9.0",
-        "puppeteer-extra": "^3.3.6",
-        "puppeteer-extra-plugin-stealth": "^2.11.2",
-        "request": "^2.88.2",
         "sharp": "^0.33.5",
         "striptags": "^3.2.0",
         "url-parse": "^1.4.3",
