--- playwright-worker/lib/chromiumBrowser.js
@@ -1,58 +1,0 @@
-const proxyChain = require('proxy-chain');
-const { chromium } = require('playwright');
-const logger = require('./logger');
-
-const BASE_ARGS = [
-  '--no-sandbox',
-  '--disable-dev-shm-usage',
-  '--disable-blink-features=AutomationControlled',
-  '--disable-features=IsolateOrigins,site-per-process',
-  '--disable-font-subpixel-positioning',
-];
-
-class ChromiumBrowser {
-  constructor(debug = false, local = false) {
-    this.debug = debug;
-    this.local = local;
-    this.browser = null;
-    this.anonymizedProxy = null;
-    this.staticArgs = BASE_ARGS;
-    this.localExecutivePath = '/usr/bin/chromium-browser';
-  }
-
-  async getBrowser(proxy) {
-    const launchArgs = [...this.staticArgs];
-
-    const launchOptions = {
-      args: launchArgs,
-      headless: true,
-      chromiumSandbox: this.local,
-      timeout: 120000,
-    };
-
-    if (proxy) {
-      this.anonymizedProxy = await proxyChain.anonymizeProxy(proxy);
-      launchArgs.push(`--proxy-server=${this.anonymizedProxy}`);
-    }
-
-    let executablePath = this.localExecutivePath;
-
-    if (executablePath) {
-      launchOptions.executablePath = executablePath;
-    }
-
-    if (this.debug) logger.debug('Launching Chromium with options', { launchOptions });
-
-    this.browser = await chromium.launch(launchOptions);
-    return this.browser;
-  }
-
-  async closeProxy() {
-    if (this.anonymizedProxy) {
-      await proxyChain.closeAnonymizedProxy(this.anonymizedProxy, true);
-      this.anonymizedProxy = null;
-    }
-  }
-}
-
-module.exports = { ChromiumBrowser }
