--- playwright-worker/package.json
@@ -1,48 +1,0 @@
-{
-    "name": "diffy-playwright-worker",
-    "version": "1.0.0",
-    "description": "Diffy local screenshots",
-    "bin": {
-        "diffy-cli": "./index.js"
-    },
-    "scripts": {
-        "start": "node index.js",
-        "build": "pkg package.json --out-path ./build/",
-        "lint": "npx eslint ."
-    },
-    "dependencies": {
-        "@aws-sdk/client-s3": "^3.637.0",
-        "@aws-sdk/client-sqs": "^3.637.0",
-        "@aws-sdk/credential-providers": "^3.637.0",
-        "bluebird": "^3.7.2",
-        "dotenv": "^17.2.3",
-        "fs-extra": "^8.1.0",
-        "iltorb": "^2.4.4",
-        "minimist": "^1.2.8",
-        "playwright": "^1.56.0",
-        "playwright-core": "^1.49.0",
-        "playwright-extra": "^4.3.6",
-        "playwright-extra-plugin-stealth": "^0.0.1",
-        "process": "^0.11.10",
-        "proxy-chain": "^2.4.0",
-        "sharp": "^0.33.5",
-        "striptags": "^3.2.0",
-        "url-parse": "^1.4.3",
-        "winston": "^3.17.0",
-        "winston-daily-rotate-file": "^5.0.0",
-        "yargs": "^15.1.0"
-    },
-    "devDependencies": {
-        "eslint": "^8.26.0",
-        "eslint-config-standard": "^17.0.0",
-        "eslint-plugin-import": "^2.26.0",
-        "eslint-plugin-n": "^15.3.0",
-        "eslint-plugin-promise": "^6.1.1"
-    },
-    "pkg": {
-        "assets": [
-            "./node_modules/lambdafs/build/iltorb/build/bindings/iltorb.node",
-            "./node_modules/iltorb/build/bindings/iltorb.node"
-        ]
-    }
-}
